requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0
joblib>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0
pydantic>=2.4.0
PyYAML>=6.0
